+++ packages/axios-extension/src/abap/adt-catalog/services/abapcdsview-service.ts
@@ -8,7 +8,7 @@ export class AbapCDSViewService extends AdtService {
     /**
      * @see AdtService.getAdtCatagory()
      */
-    private static adtCategory = {
+    private static readonly adtCategory = {
         scheme: 'http://www.sap.com/adt/categories/respository',
         term: 'search'
     };