+++ packages/fiori-app-sub-generator/test/int/fiori-freestyle/expected-output/simple_typescript/eslint.config.js
@@ -1,0 +1,5 @@
+const fioriTools  = require('@sap-ux/eslint-plugin-fiori-tools');
+
+module.exports = [
+    ...fioriTools.configs.recommended
+];