+++ packages/preview-middleware-client/test/unit/flp/enableFakeConnector.test.ts
@@ -65,7 +65,9 @@ describe('flp/FakeLrepConnector', () => {
             };
             fetchMock.mockResolvedValue({ text: jest.fn(), ok: true });
             documentMock.getElementById.mockReturnValue({
-                getAttribute: jest.fn().mockReturnValue('{"generator":"@sap-ux/control-property-editor"}')
+                dataset: {
+                    openUxPreviewFlexSettings: JSON.stringify({ generator: '@sap-ux/control-property-editor' })
+                }
             });
 
             await create([change]);
@@ -89,7 +91,9 @@ describe('flp/FakeLrepConnector', () => {
             };
             fetchMock.mockResolvedValue({ text: jest.fn(), ok: true });
             documentMock.getElementById.mockReturnValue({
-                getAttribute: jest.fn().mockReturnValue('{"generator":"@sap-ux/control-property-editor"}')
+                dataset: {
+                    openUxPreviewFlexSettings: JSON.stringify({ generator: '@sap-ux/control-property-editor' })
+                }
             });
 
             await create(change);