+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.40.3
+
+### Patch Changes
+
+- 2204ad3: fix(deps): update dependencies @sap-ux/annotation-converter to v0.10.19 and @sap-ux/vocabularies-types to v0.14.5
+- Updated dependencies [2204ad3]
+    - @sap-ux/fiori-annotation-api@0.8.3
+
+## 0.40.2
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+- Updated dependencies [c7f9a60]
+    - @sap-ux/logger@0.8.0
+    - @sap-ux/project-access@1.33.0
+    - @sap-ux/fiori-annotation-api@0.8.2
+
+## 0.40.1
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/fiori-annotation-api@0.8.1
+    - @sap-ux/project-access@1.32.17
+    - @sap-ux/logger@0.7.3
+
+## 0.40.0
+
+### Minor Changes
+
+- 5287327: Updated @sap-ux/annotation-converter to version 0.10.9 and @sap-ux/vocabularies-types to version 0.13.2 across multiple packages. These changes ensure that the latest versions with potential fixes and enhancements are used.
+
+### Patch Changes
+
+- Updated dependencies [5287327]
+    - @sap-ux/fiori-annotation-api@0.8.0
+
 ## 0.39.33
 
 ### Patch Changes