+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/project-access
 
+## 1.33.0
+
+### Minor Changes
+
+- c7f9a60: Export fioriToolsDirectory
+
+### Patch Changes
+
+- @sap-ux/ui5-config@0.29.12
+
+## 1.32.17
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/ui5-config@0.29.12
+    - @sap-ux/i18n@0.3.7
+
 ## 1.32.16
 
 ### Patch Changes