+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,111 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.78
+
+### Patch Changes
+
+- Updated dependencies [2204ad3]
+    - @sap-ux/odata-service-inquirer@2.13.9
+    - @sap-ux/axios-extension@1.25.5
+    - @sap-ux/inquirer-common@0.10.7
+    - @sap-ux/abap-deploy-config-inquirer@1.6.71
+
+## 0.2.77
+
+### Patch Changes
+
+- Updated dependencies [4e0f204]
+    - @sap-ux/axios-extension@1.25.4
+    - @sap-ux/abap-deploy-config-inquirer@1.6.70
+    - @sap-ux/odata-service-inquirer@2.13.8
+
+## 0.2.76
+
+### Patch Changes
+
+- Updated dependencies [6382440]
+    - @sap-ux/inquirer-common@0.10.6
+    - @sap-ux/abap-deploy-config-inquirer@1.6.69
+    - @sap-ux/odata-service-inquirer@2.13.7
+
+## 0.2.75
+
+### Patch Changes
+
+- @sap-ux/odata-service-inquirer@2.13.6
+- @sap-ux/abap-deploy-config-inquirer@1.6.68
+- @sap-ux/inquirer-common@0.10.5
+
+## 0.2.74
+
+### Patch Changes
+
+- @sap-ux/abap-deploy-config-inquirer@1.6.67
+- @sap-ux/inquirer-common@0.10.4
+- @sap-ux/odata-service-inquirer@2.13.5
+
+## 0.2.73
+
+### Patch Changes
+
+- Updated dependencies [14a1bc2]
+    - @sap-ux/axios-extension@1.25.3
+    - @sap-ux/abap-deploy-config-inquirer@1.6.66
+    - @sap-ux/odata-service-inquirer@2.13.4
+
+## 0.2.72
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/logger@0.8.0
+    - @sap-ux/abap-deploy-config-inquirer@1.6.65
+    - @sap-ux/axios-extension@1.25.2
+    - @sap-ux/btp-utils@1.1.6
+    - @sap-ux/inquirer-common@0.10.3
+    - @sap-ux/odata-service-inquirer@2.13.3
+    - @sap-ux/telemetry@0.6.50
+
+## 0.2.71
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/abap-deploy-config-inquirer@1.6.64
+    - @sap-ux/odata-service-inquirer@2.13.2
+    - @sap-ux/axios-extension@1.25.1
+    - @sap-ux/inquirer-common@0.10.2
+    - @sap-ux/telemetry@0.6.49
+    - @sap-ux/logger@0.7.3
+    - @sap-ux/btp-utils@1.1.6
+
+## 0.2.70
+
+### Patch Changes
+
+- Updated dependencies [cda79d4]
+    - @sap-ux/odata-service-inquirer@2.13.1
+
+## 0.2.69
+
+### Patch Changes
+
+- Updated dependencies [691ab27]
+    - @sap-ux/odata-service-inquirer@2.13.0
+    - @sap-ux/inquirer-common@0.10.1
+    - @sap-ux/abap-deploy-config-inquirer@1.6.63
+
+## 0.2.68
+
+### Patch Changes
+
+- Updated dependencies [5287327]
+    - @sap-ux/odata-service-inquirer@2.12.0
+    - @sap-ux/axios-extension@1.25.0
+    - @sap-ux/inquirer-common@0.10.0
+    - @sap-ux/abap-deploy-config-inquirer@1.6.62
+
 ## 0.2.67
 
 ### Patch Changes